+++ src/syntax/binding.rs
@@ -14,16 +14,15 @@ pub fn try_parse<'a>(tokens: &'a Vec<Token>, pos: usize) -> ParsingResult<Bindin
      * let keyword
      */
     let (is_mutable, binding_token, next_pos) = {
-        let let_token = parse_token_type(tokens, current_pos, TokenType::LET);
-        match let_token {
-            Ok((let_token, next_let)) => {
-                let mut_token = parse_token_type(tokens, next_let, TokenType::MUT);
-                match mut_token {
-                    Ok((_mut_token, next_mut)) => (true, let_token, next_mut),
-                    _ => (false, let_token, next_let),
+        match parse_token_type(tokens, current_pos, TokenType::VAL) {
+            Ok((val_token, next_pos)) => (false, val_token, next_pos),
+            _ => {
+                // If VAL is not found, search for VAR
+                match parse_token_type(tokens, current_pos, TokenType::VAR) {
+                    Ok((var_token, next_pos)) => (true, var_token, next_pos),
+                    _ => return Err(ParsingError::Unmatched),
                 }
             }
-            _ => return Err(ParsingError::Unmatched),
         }
     };
     current_pos = next_pos;
@@ -112,7 +111,7 @@ mod tests {
 
     #[test]
     fn should_parse_val_binding() {
-        let tokens = get_tokens(&String::from("let identifier = 20")).unwrap();
+        let tokens = get_tokens(&String::from("val identifier = 20")).unwrap();
         let Ok((binding, _)) = try_parse(&tokens, 0) else {
             panic!()
         };
@@ -122,11 +121,11 @@ mod tests {
 
     #[test]
     fn should_parse_val() {
-        let tokens = get_tokens(&String::from("let")).unwrap();
-        let (token, _) = parse_token_type(&tokens, 0, TokenType::LET).unwrap();
+        let tokens = get_tokens(&String::from("val")).unwrap();
+        let (token, _) = parse_token_type(&tokens, 0, TokenType::VAL).unwrap();
 
-        assert_eq!(TokenType::LET, token.token_type);
-        assert_eq!("let", token.value);
+        assert_eq!(TokenType::VAL, token.token_type);
+        assert_eq!("val", token.value);
     }
 
     #[test]
@@ -168,8 +167,8 @@ mod tests {
 
     #[test]
     fn should_return_correct_error() {
-        let tokens = get_tokens(&String::from("let")).unwrap();
-        assert_eq!(TokenType::LET, tokens[0].token_type);
+        let tokens = get_tokens(&String::from("val")).unwrap();
+        assert_eq!(TokenType::VAL, tokens[0].token_type);
         assert_eq!(0, tokens[0].position);
         let binding = try_parse(&tokens, 0);
 
@@ -184,8 +183,8 @@ mod tests {
 
     #[test]
     fn should_return_error_when_identifier_is_wrong() {
-        let tokens = get_tokens(&String::from("let 322")).unwrap();
-        assert_eq!(TokenType::LET, tokens[0].token_type);
+        let tokens = get_tokens(&String::from("val 322")).unwrap();
+        assert_eq!(TokenType::VAL, tokens[0].token_type);
         assert_eq!(0, tokens[0].position);
         let binding = try_parse(&tokens, 0);
 
@@ -197,7 +196,7 @@ mod tests {
             _ => panic!("Error expected"),
         }
 
-        let tokens = get_tokens(&String::from("let \"hello\"")).unwrap();
+        let tokens = get_tokens(&String::from("val \"hello\"")).unwrap();
         let binding = try_parse(&tokens, 0);
 
         match binding {
@@ -211,7 +210,7 @@ mod tests {
 
     #[test]
     fn should_return_error_when_equal_op_is_wrong() {
-        let tokens = get_tokens(&String::from("let id \"error\"")).unwrap();
+        let tokens = get_tokens(&String::from("val id \"error\"")).unwrap();
         let binding = try_parse(&tokens, 0);
 
         match binding {